Replace an overloaded function with a default parameter
This commit eliminates an overload of the function GetGain_1D_OutZero function.
Instead, there is now one function, with different behavior based on if the default
parameter (ant_number=0) is overridden during the function call.

This allows us to consolidate that function into one.
This is good beceause it is simpler, and leaves fewer LOC to be debugged.
The functions were nearly completely redundant anyway.
This came up because one version of the function had safeguards against negative gains
but the other didn't. But, having two version of the functions was unnecessary in the first place.
So they are consolidated, which is better code practice anyway.