Increase the Antenna_r class class def version
This increases the Antenna_r (antenna report) class definition
version number from 1 to 2.
This is important because in a previos commit
(1b10877) (f1ab79eff3e66fcf50b032ec38ba26b28a090f6a)
a new member was added to the class: `ray_step`.
This changes the class definition, so if you try to read older
AraSim files with the new class definition, you get a seg fault.
(Since the library definition of the class and the version
written to disk now clash.)
Bumping the class version seems to prevent this crash.